+++ src/Execution/Arguments/ArgumentSetFactory.php
@@ -42,7 +42,6 @@ class ArgumentSetFactory
      * Wrap client-given args with type information.
      *
      * @param  array<mixed>  $args
-     * @return \Nuwave\Lighthouse\Execution\Arguments\ArgumentSet
      */
     public function fromResolveInfo(array $args, ResolveInfo $resolveInfo): ArgumentSet
     {
@@ -61,7 +60,6 @@ class ArgumentSetFactory
      *
      * @param  \GraphQL\Language\AST\FieldDefinitionNode|\GraphQL\Language\AST\InputObjectTypeDefinitionNode  $definition
      * @param  array<mixed>  $args
-     * @return \Nuwave\Lighthouse\Execution\Arguments\ArgumentSet
      */
     public function wrapArgs(Node $definition, array $args): ArgumentSet
     {
@@ -110,7 +108,6 @@ class ArgumentSetFactory
      * Wrap a single client-given argument with type information.
      *
      * @param  mixed  $value The client given value.
-     * @return \Nuwave\Lighthouse\Execution\Arguments\Argument
      */
     protected function wrapInArgument($value, InputValueDefinitionNode $definition): Argument
     {
@@ -158,7 +155,6 @@ class ArgumentSetFactory
      * Wrap a client-given value with information from a named type.
      *
      * @param  mixed  $value The client given value.
-     * @param  \Nuwave\Lighthouse\Execution\Arguments\NamedType  $namedType
      * @return \Nuwave\Lighthouse\Execution\Arguments\ArgumentSet|mixed
      */
     protected function wrapWithNamedType($value, NamedType $namedType)