+++ src/Execution/Arguments/NestedOneToMany.php
@@ -38,9 +38,6 @@ class NestedOneToMany implements ArgResolver
         }
     }
 
-    /**
-     * @param  \Nuwave\Lighthouse\Execution\Arguments\ArgumentSet  $args
-     */
     public static function createUpdateUpsert(ArgumentSet $args, Relation $relation): void
     {
         if ($args->has('create')) {