+++ src/Federation/SchemaPrinter.php
@@ -30,7 +30,7 @@ class SchemaPrinter extends GraphQLSchemaPrinter
                         .$f->name
                         .static::printArgs($options, $f->args, '  ')
                         .': '
-                        .(string) $f->getType()
+                        .$f->getType()
                         .(isset($options['printDirectives'])
                             ? $options['printDirectives']($f)
                             : '')