+++ src/Validation/RulesGatherer.php
@@ -3,6 +3,7 @@
 namespace Nuwave\Lighthouse\Validation;
 
 use Illuminate\Contracts\Validation\Rule;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Collection;
 use Illuminate\Validation\ValidationRuleParser;
 use Nuwave\Lighthouse\Execution\Arguments\ArgumentSet;
@@ -11,8 +12,10 @@ use Nuwave\Lighthouse\Support\Contracts\ArgDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgDirectiveForArray;
 use Nuwave\Lighthouse\Support\Contracts\ArgumentSetValidation;
 use Nuwave\Lighthouse\Support\Contracts\ArgumentValidation;
+use Nuwave\Lighthouse\Support\Contracts\WithReferenceRule;
 use Nuwave\Lighthouse\Support\Traits\HasArgumentValue;
 use Nuwave\Lighthouse\Support\Utils;
+use Throwable;
 
 class RulesGatherer
 {
@@ -222,6 +225,10 @@ class RulesGatherer
         return array_map(
             static function ($rule) use ($argumentPath) {
                 if (is_object($rule)) {
+                    if ($rule instanceof WithReferenceRule) {
+                        $rule->setArgumentPath($argumentPath);
+                    }
+
                     return $rule;
                 }
 
@@ -235,6 +242,21 @@ class RulesGatherer
 
                 $name = $parsed[0];
                 $args = $parsed[1];
+
+                if ($name === 'WithReference') {
+                    $indexes = explode('_', $args[1]);
+                    array_splice($args, 1, 1);
+                    foreach ($indexes as $index) {
+                        // Skipping over the first index, which is the name
+                        $index = (int) $index + 1;
+                        $args[$index] = implode('.', array_merge($argumentPath, [$args[$index]]));
+                    }
+
+                    $parsed = ValidationRuleParser::parse($args);
+
+                    $name = $parsed[0];
+                    $args = $parsed[1];
+                }
 
                 // Those rule lists are a subset of https://github.com/illuminate/validation/blob/8079fd53dee983e7c52d1819ae3b98c71a64fbc0/Validator.php#L206-L236
                 // using the docs to know which ones reference other fields: https://laravel.com/docs/8.x/validation#available-validation-rules
@@ -269,6 +291,20 @@ class RulesGatherer
                         },
                         $args
                     );
+                }
+
+                // Rules where the first argument is a date or a field reference
+                if (is_string($args[0] ?? null) && in_array($name, [
+                    'After',
+                    'AfterOrEqual',
+                    'Before',
+                    'BeforeOrEqual',
+                ])) {
+                    try {
+                        Carbon::parse($args[0]);
+                    } catch (Throwable $argumentIsNotADate) {
+                        $args[0] = implode('.', array_merge($argumentPath, [$args[0]]));
+                    }
                 }
 
                 // Laravel expects the rule to be a flat array of name, arg1, arg2, ...